discogs_credits: extract preflight resolvers + hasDiscogsLinkDefined
Issue #32 — two more moves out of the entry:

  src/preflight.js
    checkMissingArtists, checkMissingCompanies. The two near-duplicate
    "IDB → name search → URL relation" lookup loops that run before
    the review table. The ANALYSIS §2.3-mandated collapse into a single
    resolveEntity is a separate later chunk; this one is just-move,
    no-behavior-change.

  src/api-mb.js (extended)
    hasDiscogsLinkDefined — the one-shot /ws/js/release/<mbid>?inc=rels
    probe used at import start to find the Discogs URL linked to a
    release. Skips mbThrottle deliberately (one-shot, not in hot loop).

~360 lines moved out. Full 12-fixture suite still green.
